Update: Routing of both versions of map
The initial version where maps are rendered using vanilla js is now at /dev and the new version which uses react components is at /
